pyproject.toml
[project]
name = "salesforce-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.17.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]
